requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
brotli==1.0.9
django-cors-headers==3.11.0
django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
setuptools==62.0.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
whitenoise==6.0.0
cffi==1.15.0
django-dotenv==1.4.2
djangorestframework-simplejwt==5.1.0
django-filter==22.1
pycparser==2.21
PyJWT==1.7.1
tzdata==2022.1
Pillow==9.1.0